+++ Makefile
@@ -1,11 +1,11 @@
 .PHONY: copyright-check apply-copyright fix-licenses check-licenses
 ## Copyright checks
 copyright-check:
-	docker run -it --rm -v $(CURDIR):/github/workspace apache/skywalking-eyes header check
+	docker run -it --rm -v $(CURDIR):/github/workspace apache/skywalking-eyes -c .github/.licenserc.yaml header check
 
 ## Add copyright notice to new files
 apply-copyright:
-	docker run -it --rm -v $(CURDIR):/github/workspace apache/skywalking-eyes header fix
+	docker run -it --rm -v $(CURDIR):/github/workspace apache/skywalking-eyes -c .github/.licenserc.yaml header fix
 
 fix-licenses: apply-copyright
