+++ Pulumi.yaml
@@ -6,3 +6,4 @@ config:
   pulumi:tags:
     value:
       pulumi:template: python
+  datarobot:tracecontext: guarded-rag-assistant
