+++ infra/feature_flag_requirements.yaml
@@ -7,3 +7,4 @@ ENABLE_CUSTOM_MODEL_PREDICT_RESPONSE_EXTRA_MODEL_OUTPUT: true
 ENABLE_MODERATION_GUARDRAILS: true
 ENABLE_MMM_DATA_QUALITY: true
 DISABLE_CUSTOM_APPS_WORKSHOP: false
+ENABLE_GENAI_EXPERIMENTATION: true
