+++ infra/settings_grader.py
@@ -50,9 +50,7 @@ deployment_args = DeploymentArgs(
     predictions_settings=(
         None
         if default_prediction_server_id
-        else datarobot.DeploymentPredictionsSettingsArgs(
-            min_computes=0, max_computes=1, real_time=True
-        )
+        else datarobot.DeploymentPredictionsSettingsArgs(min_computes=0, max_computes=1)
     ),
     predictions_data_collection_settings=datarobot.DeploymentPredictionsDataCollectionSettingsArgs(
         enabled=True
