--- mypy.ini
@@ -1,14 +1,0 @@
-[mypy]
-plugins = pydantic.mypy
-strict = true
-ignore_missing_imports = true
-exclude = (?x)(
-    output/ # exclude files starting with "output"
-    | deployment_.*/
-    | include/   # or files starting with "include."
-    | venv/
-    | .venv/
-    | .cache
-    | tests/
-    | e2e/
-  )
