--- pytest.ini
@@ -1,2 +1,0 @@
-[pytest]
-norecursedirs=tests/e2e
