+++ requirements.txt
@@ -1,5 +1,5 @@
 pulumi>=3.133.0,<4
-pulumi-datarobot>=0.2.5,<0.3
+pulumi-datarobot>=0.4.5,<0.5
 
 datarobot>=3.5.2,<3.6
 datarobot-predict>=1.9.6,<1.10
